requirements.txt
# Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Survival Analysis
lifelines>=0.27.0
scikit-survival>=0.17.0

# Machine Learning
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0

# Deep Learning (optional)
# tensorflow>=2.8.0
# torch>=1.10.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Automated EDA
ydata-profiling>=4.0.0

# Statistical Analysis
scipy>=1.7.0
statsmodels>=0.13.0

# Utilities
pyyaml>=5.4.0
joblib>=1.0.0
tqdm>=4.62.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
